add dockerfile + fix db

+++ internal/pgstore/migrations/20221018131500-wallet.sql
@@ -3,8 +3,8 @@ CREATE TABLE IF NOT EXISTS wallet
 (
     id               bigserial PRIMARY KEY                  NOT NULL,
     owner_id         int UNIQUE                             NOT NULL,
-    balance          float                                  NOT NULL,
-    reserved_balance float                                  NOT NULL,
+    balance          numeric(11,2)                          NOT NULL,
+    reserved_balance numeric(11,2)                          NOT NULL,
     created_at       timestamp with time zone DEFAULT NOW() NOT NULL,
     updated_at       timestamp with time zone DEFAULT NOW() NOT NULL
     );
@@ -19,7 +19,7 @@ CREATE TABLE transaction
 (
     id               bigserial PRIMARY KEY                  NOT NULL,
     wallet_id        bigint REFERENCES wallet (id)          NOT NULL,
-    amount           float                                  NOT NULL,
+    amount           numeric(11,2)                          NOT NULL,
     target_wallet_id bigint REFERENCES wallet (id),
     service_id       int REFERENCES services (id),
     comment          text                                   NOT NULL,
@@ -32,7 +32,7 @@ CREATE TABLE reserved_funds
     order_id   int UNIQUE                             NOT NULL,
     owner_id   int REFERENCES wallet (owner_id)       NOT NULL,
     service_id int REFERENCES services (id)           NOT NULL,
-    amount     float                                  NOT NULL,
+    amount     numeric(11,2)                          NOT NULL,
     status     text                                   NOT NULL,
     created_at timestamp with time zone DEFAULT NOW() NOT NULL,
     updated_at timestamp with time zone DEFAULT NOW() NOT NULL
